requirements.txt
flask
boto3==1.4.4
pyyaml
python-json-logger==0.1.7
exifread
pytest
urllib3
requests
pathlib
